:hammer: 제네릭을 활용한 any 타입 제거 & 의존성 주입
